Make magic energy an int and adjust logic
Convert CEMagicEnergyContainerComponent Energy and MaxEnergy from FixedPoint2 to int and update affected logic across client, server, and shared systems. Adjust UI and price calculations to remove unnecessary casts, change ChangeMaximumEnergy signature to accept int, and ensure energy clamping/delta math uses integer types. This simplifies numeric handling for magic energy and removes redundant float/double conversions.

+++ Content.Client/_CE/Mana/CEMagicEnergySystem.cs
@@ -59,9 +59,9 @@ public sealed class CEMagicEnergyStatusControl : Control
     {
         base.FrameUpdate(args);
 
-        _progress.Value = (float)(_parent.Comp.Energy / _parent.Comp.MaxEnergy);
+        _progress.Value = _parent.Comp.Energy / _parent.Comp.MaxEnergy;
 
-        var power = (int) (_parent.Comp.Energy / _parent.Comp.MaxEnergy * 100);
+        var power = _parent.Comp.Energy / _parent.Comp.MaxEnergy * 100;
         _label.Text = $"{power}%";
     }
 }
